+++ src/onegov.core/HISTORY.rst
@@ -1,6 +1,9 @@
 Changelog
 ---------
 
+- Adds the model to the form object created by the form directive.
+  [href]
+
 - Adds the ability to recompile themes using shift+f5 in the browser.
   This option has to be enabled using the 'allow_shift_f5_compile' flag.
   [href]